Created the Database and more
I created a query to construct the database.  It will create the
database, all of the tables, and populate almost all of the tables with
data. I also cleaned up some asset files and created a mainAsset folder
to hold the game photos. Also, I connected each person to their avatar.